[PersistentNodeKeys] Refactoring: moved AsymmetricCipherKeyPair from App
